% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlphaCol.R
\name{addAlphaCol}
\alias{addAlphaCol}
\title{Add transparency to color.}
\usage{
addAlphaCol(cols, alpha)
}
\arguments{
\item{cols}{Colors in hex format or R color code e.g. "red", "black", etc.}

\item{alpha}{Alpha value.}
}
\value{
Colors with alpha value in hex format.
}
\description{
Add transparency to color.
}
