% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildControl.R
\name{buildControl}
\alias{buildControl}
\title{Build control regions}
\usage{
buildControl(
  case,
  ctrl.rel.pos,
  genome,
  output.path = "control/",
  verbose = TRUE
)
}
\value{
Control in Coordinate class object format.
}
\description{
Build control regions
}
\section{Fields}{

\describe{
\item{\code{case}}{Case in Coordinate class object format.}

\item{\code{ctrl.rel.pos}}{Control reltive position.}

\item{\code{genome}}{Genome class object.}

\item{\code{output.path}}{Output directory path to save control coordinate.}
}}

