% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKmerTable.R
\name{buildKmerTable}
\alias{buildKmerTable}
\title{Count k-mers from given sequence(s) and build a data.table of k-mer counts.}
\usage{
buildKmerTable(dna.seqs, k, method = "auto", remove.N = TRUE)
}
\arguments{
\item{dna.seqs}{String of sequence(s).}

\item{k}{Size of kmer.}

\item{method}{K-mer counting method: "Biostrings", "sliding", or "auto".
Default is "auto"; For k > 8, sliding method is used.}

\item{remove.N}{Remove unknown base? Default is TRUE.}
}
\value{
A \code{data.table} object with column kmer and N.
}
\description{
Only existed k-mers are returned in data.table object.
}
