% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMidPatternKmerTable.R
\name{buildMidPatternKmerTable}
\alias{buildMidPatternKmerTable}
\title{Count k-mers with specified middle pattern from given sequence(s) and build
a data.table of k-mer counts.}
\usage{
buildMidPatternKmerTable(dna.seqs, k, mid.patterns, remove.N = TRUE)
}
\arguments{
\item{dna.seqs}{String of sequence(s).}

\item{k}{Size of kmer.}

\item{mid.patterns}{Middle patterns.}

\item{remove.N}{Remove unknown base? Default is TRUE.}
}
\value{
A \code{data.table} object with column kmer and N.
}
\description{
Only existed k-mers are returned in data.table object.
}
