% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRESTurl.R
\name{buildRESTurl}
\alias{buildRESTurl}
\title{Function constructs a URL for a REST API call by appending query parameters.}
\usage{
buildRESTurl(url, .list = list(), ...)
}
\arguments{
\item{url}{Base URL of the REST API.}

\item{.list}{A list of named query parameters.}

\item{...}{additional optional arguments}
}
\value{
string of the full REST API URL.
}
\description{
Function constructs a URL for a REST API call by appending query parameters.
}
