% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calKmerSkew.R
\name{calKmerSkew}
\alias{calKmerSkew}
\title{Function calculates the skew of k-mers based on their occurrence in
positive and negative strands.}
\usage{
calKmerSkew(kmer.table)
}
\arguments{
\item{kmer.table}{data.table with columns: kmer, pos_strand, neg_strand.}
}
\value{
data.table with the kmer_skew column.
}
\description{
Function calculates the skew of k-mers based on their occurrence in
positive and negative strands.
}
