% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catHeader.R
\name{catHeader}
\alias{catHeader}
\title{Function prints a given message in a formatted header with borders.}
\usage{
catHeader(msg)
}
\arguments{
\item{msg}{message to be printed within the header.}
}
\description{
Function prints a given message in a formatted header with borders.
}
