% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_substring_regex}
\alias{count_substring_regex}
\title{Count sequence content in a given sequence.}
\usage{
count_substring_regex(sequence, start, end, pattern)
}
\arguments{
\item{sequence}{A sequence to map.}

\item{start}{Start positions.}

\item{end}{End positions.}

\item{pattern}{A regex pattern to search for within start and end positions.}
}
\value{
A numeric vector of count.
}
\description{
stringi has no function that search within substring without memory creating
it. This function solve that. Unlike count_substring_fixed, this function
does not need to memory copy substring.
}
