% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBaseComposition.R
\name{countBaseComposition}
\alias{countBaseComposition}
\title{Function performs an analysis of base composition including sequence
frequency, PWM calculations, and G/C content at various window sizes.}
\usage{
countBaseComposition(case, genome, case.pattern, output.path = "./")
}
\arguments{
\item{case}{A Coordinate class object or similar structure.}

\item{genome}{Genome class object or similar structure.}

\item{case.pattern}{String patterns to consider in the analysis.}

\item{output.path}{Output path for saving the analysis results.}
}
\description{
Function performs an analysis of base composition including sequence
frequency, PWM calculations, and G/C content at various window sizes.
}
