% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDistribution.R
\name{countDistribution}
\alias{countDistribution}
\title{Function performs an analysis of the distribution of genomic cases.}
\usage{
countDistribution(case, genome, case.pattern, output.path = "./")
}
\arguments{
\item{case}{A Coordinate class object or similar structure for genomic data.}

\item{genome}{Genome class object or similar structure.}

\item{case.pattern}{String patterns to consider in the analysis.}

\item{output.path}{Output path for saving the analysis results.}
}
\description{
Check case distribution in replicates, chromosomes, and strands.
Check case base composition and filter out other case.patterns.
Then, it generates various plots like bar plots and Venn/Euler diagrams.
}
