% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countKmers}
\alias{countKmers}
\title{Count k-mers from string(s) using a simple hash table.}
\usage{
countKmers(sequences, k)
}
\arguments{
\item{sequences}{Sequence strings.}

\item{k}{Size of k-mer.}
}
\value{
A vector of k-mer counts. The counts of multiple sequences are
combined, similar to Biostrings::oligonucleotideFrequency simplify.as
"collapsed".
}
\description{
Count only observed k-mers. Biostrings::oligonucleotideFrequency reports all
possible k-mers. For k > 12, the memory for creating empty k-mer counts
spiked and crashed the R session.
}
