% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countMidPatternContext}
\alias{countMidPatternContext}
\title{Locate a middle sequence pattern and count its sequence context.}
\usage{
countMidPatternContext(sequence, mid_pattern, window, context_pattern)
}
\arguments{
\item{sequence}{A sequence to slide.}

\item{mid_pattern}{A middle pattern to search for.}

\item{window}{Size of a surrounding window.}

\item{context_pattern}{A context pattern to search for.}
}
\value{
A numeric vector of count.
}
\description{
Locate a middle sequence pattern and count its sequence context.
}
