% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countMidPatternRangedKmers}
\alias{countMidPatternRangedKmers}
\title{Count k-mers in given ranges of a sequence.}
\usage{
countMidPatternRangedKmers(sequence, starts, ends, k, mid_pattern)
}
\arguments{
\item{sequence}{A sequence to count.}

\item{starts}{Start positions.}

\item{ends}{End positions.}

\item{k}{K-mer size.}

\item{mid_pattern}{A middle pattern to search for.}
}
\value{
A k-mer-named vector of count.
}
\description{
Slide and update the cummulated table count.
}
