% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countPointContext}
\alias{countPointContext}
\title{Ccount sequence context of given point positions.}
\usage{
countPointContext(sequence, points, len, window, context_pattern)
}
\arguments{
\item{sequence}{A sequence to slide.}

\item{points}{Middle point positions.}

\item{len}{Length of the middle point. Default to 1.}

\item{window}{Size of a surrounding window.}

\item{context_pattern}{A context pattern to search for.}
}
\value{
A numeric vector of count.
}
\description{
Ccount sequence context of given point positions.
}
