% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countPointContext2}
\alias{countPointContext2}
\title{Ccount sequence context of given point positions.}
\usage{
countPointContext2(sequence, points, len, window, context_patterns)
}
\arguments{
\item{sequence}{A sequence to slide.}

\item{points}{Middle point positions.}

\item{len}{Length of the middle point.}

\item{window}{Size of a surrounding window.}

\item{context_patterns}{Context patterns to search for.}
}
\value{
A named vector of frequency of counts.
}
\description{
Ccount sequence context of given point positions.
}
