% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countRangedKmers}
\alias{countRangedKmers}
\title{Count k-mers in given ranges of a sequence.}
\usage{
countRangedKmers(sequence, starts, ends, k)
}
\arguments{
\item{sequence}{A sequence to count.}

\item{starts}{Start positions.}

\item{ends}{End positions.}

\item{k}{K-mer size.}
}
\value{
A k-mer-named vector of count.
}
\description{
Slide and update the cummulated table count.
}
