% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRevCompKmers.R
\name{countRevCompKmers}
\alias{countRevCompKmers}
\title{Count reverse complement sequence from its opposite strand.
Build for k-mer table generated from initKmerTable function but applicable to
others with the same format.}
\usage{
countRevCompKmers(kmer.table)
}
\arguments{
\item{kmer.table}{A data.table of k-mer with at least 3 columns: kmer,
pos_strand, and neg_strand. Splitted k-mer columns: kmer_part1 and
kmer_part2 is supported.}
}
\value{
Updated k-mer table.
}
\description{
Count reverse complement sequence from its opposite strand.
Build for k-mer table generated from initKmerTable function but applicable to
others with the same format.
}
