% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countSlidingWindow}
\alias{countSlidingWindow}
\title{Count sequence content in a sliding window of a sequence.}
\usage{
countSlidingWindow(sequence, window, pattern)
}
\arguments{
\item{sequence}{A sequence to slide.}

\item{window}{Size of a window.}

\item{pattern}{A pattern to search for.}
}
\value{
A numeric vector of count.
}
\description{
Count sequence content in a sliding window of a sequence.
}
