% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countSlidingWindow2}
\alias{countSlidingWindow2}
\title{Count sequence content in a sliding window of a sequence.}
\usage{
countSlidingWindow2(sequence, window, patterns)
}
\arguments{
\item{sequence}{A sequence to slide.}

\item{window}{Size of a window.}

\item{patterns}{Patterns of the same size to search for.}
}
\value{
Named vector of frequency of count.
}
\description{
Count sequence content in a sliding window of a sequence.
}
