% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOSMICauthURL.R
\name{getCOSMICauthURL}
\alias{getCOSMICauthURL}
\title{Get COSMIC authenticated URL.}
\usage{
getCOSMICauthURL(email, password, url)
}
\arguments{
\item{email}{Email registered with COSMIC.}

\item{password}{Password associated with the registered email.}

\item{url}{Public URL provided by the COSMIC website for data access.}
}
\value{
Authenticated URL valid for 1-hour access to COSMIC data.
}
\description{
To access the data for non-commercial usage, you must register with the
COSMIC. This function fetch the authenticated URL from the public URL
given by the COSMIC website.
}
