% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOSMICcancerGeneCensus.R
\name{getCOSMICcancerGeneCensus}
\alias{getCOSMICcancerGeneCensus}
\title{Get Cancer Gene Census (CGC) from COSMIC database.}
\usage{
getCOSMICcancerGeneCensus(email, password)
}
\arguments{
\item{email}{Email registered with COSMIC.}

\item{password}{Password associated with the registered email.}
}
\value{
A \code{data.table} containing the Cancer Gene Census data.
}
\description{
To access the data for non-commercial usage, you must register with the
COSMIC. This function fetch the latest CGC.
}
