% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOSMIClatestVersion.R
\name{getCOSMIClatestVersion}
\alias{getCOSMIClatestVersion}
\title{Function retrieves the latest version information of the COSMIC database
and the associated genome version by scraping data from the COSMIC website.}
\usage{
getCOSMIClatestVersion()
}
\value{
A named vector containing the latest COSMIC version (\code{cosmic}) and
genome version (\code{genome}).
}
\description{
Function retrieves the latest version information of the COSMIC database
and the associated genome version by scraping data from the COSMIC website.
}
