% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOSMICmutantExport.R
\name{getCOSMICmutantExport}
\alias{getCOSMICmutantExport}
\title{Function downloads the latest Cosmic Mutant Export data from the COSMIC
database. It requires the user to be registered with COSMIC for non-commercial
use. The function constructs the URL for the latest mutant export file,
authenticates the URL, and then downloads the data.}
\usage{
getCOSMICmutantExport(email, password)
}
\arguments{
\item{email}{Email registered with COSMIC for accessing data.}

\item{password}{Password for the COSMIC account.}
}
\value{
A \code{data.table} containing the Cosmic Mutant Export data.
}
\description{
Function downloads the latest Cosmic Mutant Export data from the COSMIC
database. It requires the user to be registered with COSMIC for non-commercial
use. The function constructs the URL for the latest mutant export file,
authenticates the URL, and then downloads the data.
}
