% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsemblData.R
\name{getEnsemblData}
\alias{getEnsemblData}
\title{A generic function to get Ensembl data persistently from a URL. This is an
internal function used by other getEnsemblXXX functions.}
\usage{
getEnsemblData(url, handle, max.attempt = 5)
}
\arguments{
\item{url}{Pre-built Ensembl REST API URL.}

\item{handle}{\code{curl} handle object configured for the Ensembl REST API.}

\item{max.attempt}{Maximum number of attempts to fetch the data, default is 5.}
}
\value{
Parsed JSON data, which could be in the form of a data.frame or
a list of lists, depending on the API response.
}
\description{
Error is handled based on their rule as set out at
https://github.com/Ensembl/ensembl-rest/wiki/HTTP-Response-Codes
}
