% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsemblVariantFeatures.R
\name{getEnsemblVariantFeatures}
\alias{getEnsemblVariantFeatures}
\title{Get features of given variant IDs.}
\usage{
getEnsemblVariantFeatures(
  species,
  variant.ids,
  include.genotypes = FALSE,
  include.phenotypes = FALSE,
  include.allele.frequencies = FALSE,
  include.genotype.frequencies = FALSE,
  curl.max.con = 100,
  verbose = 1
)
}
\arguments{
\item{species}{Species name or alias e.g. homo_sapiens, human, etc.}

\item{variant.ids}{A vector of variant IDs. e.g. rs56116432, COSM476, etc.}

\item{include.genotypes}{Include genotypes? Default is FALSE.}

\item{include.phenotypes}{Include phenotypes? Default is FALSE.}

\item{include.allele.frequencies}{Include allele frequencies? Default is
FALSE.}

\item{include.genotype.frequencies}{Include genotype frequencies? Default is
FALSE.}

\item{curl.max.con}{Curl max connection. Default is 100.}

\item{verbose}{Option: 1 for showing error only and 2 for showing every
request number made. Default is 1.}
}
\value{
A variant-named list of list of variation features.
}
\description{
POST variation/:species
}
