% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICTVvirusMetadataResource.R
\name{getICTVvirusMetadataResource}
\alias{getICTVvirusMetadataResource}
\title{Get Virus Metadata Resource (VMR) from International Committee on Taxonomy
of Viruses (ICTV)}
\usage{
getICTVvirusMetadataResource()
}
\value{
Virus Metadata Resource data.table.
}
\description{
Always get the latest VMR table, so no argument.
}
