% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCBIassemblySummary.R
\name{getNCBIassemblySummary}
\alias{getNCBIassemblySummary}
\title{Get NCBI assembly summary.}
\usage{
getNCBIassemblySummary(organism.group, db = "refseq")
}
\arguments{
\item{db}{Database record to use: refseq or genbank}

\item{taxonomic.group}{Options are archaea, bacteria, fungi, invertebrate,
plant, protozoa, vertebrate_mammalian, vertebrate_other, viral, or all.}
}
\value{
Assembly summary data.table.
}
\description{
Get NCBI assembly summary.
}
