% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCBItaxonomy.R
\name{getNCBItaxonomy}
\alias{getNCBItaxonomy}
\title{Get all taxonomy classification from NCBI}
\usage{
getNCBItaxonomy()
}
\value{
A data.table of taxonomy classification.
}
\description{
NCBI provides taxonomy for all life domains in rankedlineage.dmp file.
This function download and store the file in kmertone_data path.
}
