% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScores.R
\name{getScores}
\alias{getScores}
\title{Function calculates scores for k-mers based on case and control k-mer counts.}
\usage{
getScores(case.kmers, control.kmers)
}
\arguments{
\item{case.kmers}{A data.table containing k-mer counts in case samples.}

\item{control.kmers}{A data.table containing k-mer counts in control samples.}
}
\value{
A data.table containing scores for each k-mer.
}
\description{
Function calculates scores for k-mers based on case and control k-mer counts.
}
