% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVCFmetainfo.R
\name{getVCFmetainfo}
\alias{getVCFmetainfo}
\title{Read VCF metainfo file using tabix.}
\usage{
getVCFmetainfo(vcf.file)
}
\arguments{
\item{vcf.file}{A path to a local or remote tabix-indexed VCF file.}
}
\value{
VCF metainfo.
}
\description{
Require tabix in PATH
VCF manual is referred from https://samtools.github.io/hts-specs/VCFv4.3.pdf
}
