% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomputeGenomicContents.R
\name{loadGenomicContents}
\alias{loadGenomicContents}
\title{Function calculates various genomic content metrics based on the provided genome object.}
\usage{
loadGenomicContents(genome)
}
\arguments{
\item{genome}{An object of class 'NCBI_Genome' containing genomic information.}
}
\value{
A data.table containing calculated genomic content metrics.
}
\description{
Function calculates various genomic content metrics based on the provided genome object.
}
