% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCoordinate.R
\name{mergeCoordinate}
\alias{mergeCoordinate}
\title{Merge overlapping or continuous regions.}
\usage{
mergeCoordinate(coor)
}
\value{
Merged coordinate `data.table`.
}
\description{
Table must have start and end columns. The output is exactly similar to the
   reduce function from GenomicRanges.
}
\section{Fields}{

\describe{
\item{\code{coor}}{Coordinate `data.table`.}
}}

