% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixColors.R
\name{mixColors}
\alias{mixColors}
\title{Mix color}
\usage{
mixColors(cols, alpha)
}
\arguments{
\item{cols}{Colors in hex format or R color code e.g. "red", "black", etc.}

\item{alpha}{Add alpha transparency value.}
}
\value{
New mixed colors in hex format.
}
\description{
This is useful to get overlayed colors.
}
