% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistentDownload.R
\name{persistentDownload}
\alias{persistentDownload}
\title{Download file until successful}
\usage{
persistentDownload(
  file.url,
  output.name,
  max.attempt = 5,
  user.invoke = TRUE,
  header
)
}
\arguments{
\item{file.url}{File uniform resource locator.}

\item{output.name}{Output name.}

\item{max.attempt}{Maximum number of attempt. Default is 5.}

\item{user.invoke}{If number of attempt is reached, ask user whether to keep
trying. Default is TRUE to invoke the prompt.}

\item{header}{A named list or vector of curl header.}
}
\value{
A downloaded file.
}
\description{
If download failed, it will be repeated until max attempt reached.
}
