% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBED.R
\name{readBED}
\alias{readBED}
\title{Read a BED file. One-based indexing is enforced.}
\usage{
readBED(bed.path)
}
\arguments{
\item{bed.path}{A path to a BED file.}
}
\value{
data.table.
}
\description{
Read a BED file. One-based indexing is enforced.
}
