% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFASTA.R
\name{readFASTA}
\alias{readFASTA}
\title{Read FASTA files.}
\usage{
readFASTA(fasta.file)
}
\arguments{
\item{fasta.file}{A path to a FASTA file.}
}
\value{
A sequence vector with header names
}
\description{
Read FASTA files.
}
