% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readSingleFASTA}
\alias{readSingleFASTA}
\title{Read fast single-header FASTA file.}
\usage{
readSingleFASTA(file_path, mask = "none")
}
\arguments{
\item{file_path}{A path to FASTA file.}

\item{unmask}{Capitalize all base letters?}
}
\value{
A single sequence string without header.
}
\description{
Read fast single-header FASTA file.
}
