% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF2.R
\name{readVCF2}
\alias{readVCF2}
\title{Read VCF file using tabix.}
\usage{
readVCF2(vcf.file, chr.names, starts, ends, INFO.filter = NULL)
}
\arguments{
\item{vcf.file}{A path to a local or remote tabix-indexed VCF file.}

\item{chr.names}{Chromosome names.}

\item{starts}{Start positions.}

\item{ends}{End positions.}

\item{INFO.filter}{Parse only filtered INFO ID. Default is to parse all IDs.}
}
\value{
A data.table of VCF.
}
\description{
Require tabix in PATH
VCF manual is referred from https://samtools.github.io/hts-specs/VCFv4.3.pdf
}
