% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseComplement.R
\name{reverseComplement}
\alias{reverseComplement}
\title{Get reverse complement sequence of DNA}
\usage{
reverseComplement(DNA.sequence, form = "string")
}
\arguments{
\item{DNA.sequence}{DNA sequence can be in a form of character vector or
string. Multiple sequences are accepted.}

\item{form}{Specify the form: "string" of "vector". Default is "string"}
}
\value{
Reverse complementary sequence
}
\description{
Get reverse complement sequence of DNA
}
\examples{
\donttest{
   reverseComplement("AAAAA")
   reverseComplement(c("AAAAA", "CCCCC"))
   reverseComplement(c("A", "A", "A", "A"), form = "vector")
 }
}
