% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreKmers.R
\name{scoreKmers}
\alias{scoreKmers}
\title{Function calculates the Z-score for each k-mer based on the observed
case counts and expected case counts under the null hypothesis.}
\usage{
scoreKmers(kmer.table)
}
\arguments{
\item{kmer.table}{A data.table containing k-mer counts, where each row represents a
k-mer and columns "case" and "control" represent the counts in case and control samples respectively.}
}
\value{
A modified version of the input \code{kmer.table} with an additional column
"z" containing the calculated Z-scores for each k-mer.
}
\description{
Function calculates the Z-score for each k-mer based on the observed
case counts and expected case counts under the null hypothesis.
}
