% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRepresentativeFromASM.R
\name{selectRepresentativeFromASM}
\alias{selectRepresentativeFromASM}
\title{Select the best representative species from the NCBI assembly summary.}
\usage{
selectRepresentativeFromASM(asm)
}
\arguments{
\item{asm}{NCBI assembly summary.}
}
\value{
Trimmed NCBI assembly summary.
}
\description{
sort.idx is a weight to sort where heavier will be preffered. Any tie weight
will be further sorted by organism_name. Only the top unique species_taxid
will be retained in the final assembly summary.
}
