% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFASTA.R
\name{splitFASTA}
\alias{splitFASTA}
\title{Split a FASTA file by header.}
\usage{
splitFASTA(fasta.file, output.dir = "./")
}
\arguments{
\item{fasta.file}{A path to a FASTA file.}

\item{output.dir}{A path to save the output results. Default is current working directory.}
}
\value{
A splitted fasta files by its headers.
}
\description{
The first non-space word in the header will be used as the file name.
}
\details{
data.table::fread is not built for reading in chunks. The developers expect
skip and nrow arguments to be in a small number. Large number slows the
reading a bit.
}
