% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system3.R
\name{system3}
\alias{system3}
\title{A system2 wrapper. If anything happen, just give me error!}
\usage{
system3(
  command,
  args = character(),
  stdout = "",
  stderr = "",
  stdin = "",
  input = NULL,
  env = character(),
  wait = TRUE,
  minimized,
  invisible,
  timeout = 0
)
}
\arguments{
\item{command}{the system command to be invoked, as a character string.}

\item{args}{a character vector of arguments to `command`.}

\item{stdout, stderr}{where output to ‘stdout’ or ‘stderr’ should be sent.
Possible values are "", to the R console (the default), `NULL` or `FALSE`
(discard output), `TRUE` (capture the output in a character vector) or a
character string naming a file.}

\item{stdin}{should input be diverted? "" means the default, alternatively a
character string naming a file. Ignored if input is supplied.}

\item{input}{if a character vector is supplied, this is copied one string per
line to a temporary file, and the standard input of command is redirected
to the file.}

\item{env}{character vector of name=value strings to set environment
variables.}

\item{wait}{a logical (not `NA`) indicating whether the `R` interpreter
should wait for the command to finish, or run it asynchronously. This
will be ignored (and the interpreter will always wait) if stdout = `TRUE`
or stderr = `TRUE.` When running the command asynchronously, no output
will be displayed on the Rgui console in Windows (it will be dropped,
instead).}

\item{minimized, invisible}{arguments that are accepted on Windows but ignored
on this platform, with a warning.}

\item{timeout}{timeout in seconds, ignored if 0. This is a limit for the
elapsed time running command in a separate process. Fractions of seconds
are ignored.}
}
\description{
Turn warning to error.
}
