% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimCoordinate.R
\name{trimCoordinate}
\alias{trimCoordinate}
\title{Trim out-of-bound coordinates}
\usage{
trimCoordinate(coor, seq.len, genome)
}
\value{
Trimmed coordinate `data.table`.
}
\description{
It operates in two mode: coordinate table with and without chromosome. The
   former require `Genome` for the chromosomal sequence length.
}
\section{Fields}{

\describe{
\item{\code{coor}}{Coordinate `data.table`.}

\item{\code{seq.len}}{Sequence length to trim end position.}

\item{\code{genome}}{`Genome` class object.}
}}

