% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBED.R
\name{writeBED}
\alias{writeBED}
\title{Write a BED file. Zero-based indexing is enforced.}
\usage{
writeBED(bed, output.filename)
}
\arguments{
\item{bed}{A BED \code{data.table}.}

\item{output.filename}{An output BED filename.}
}
\description{
Write a BED file. Zero-based indexing is enforced.
}
