% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFASTA.R
\name{writeFASTA}
\alias{writeFASTA}
\title{Write FASTA files.}
\usage{
writeFASTA(seqs, fasta.path, append = FALSE)
}
\arguments{
\item{seqs}{A vector or list of sequences with header name. If it is a list,
it must only contain one single sequence string for every element e.g.
list(chr1 = "NNNNNNNN") not list(chr1 = c("NNNNNN", "AAAAAA"))}

\item{out.path}{A path to a FASTA file to be saved.}
}
\value{
None
}
\description{
Write FASTA files.
}
