% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVCF.R
\name{writeVCF}
\alias{writeVCF}
\title{Write VCF file and compress using bgzip.}
\usage{
writeVCF(vcf, output.vcf.gz, append = FALSE, tabix = FALSE)
}
\arguments{
\item{vcf}{A VCF object.}

\item{output.vcf.gz}{Output filename including vcf.gz extension.}

\item{append}{To append or not? Default is FALSE.}

\item{tabix}{To tabix or not? Default is FALSE.}
}
\description{
Require bgzip in PATH
VCF manual is referred from https://samtools.github.io/hts-specs/VCFv4.3.pdf
}
