% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildControl.R
\name{buildControl}
\alias{buildControl}
\title{Build control regions}
\usage{
buildControl(
  case,
  k,
  ctrl.rel.pos,
  genome,
  output.path = "control/",
  verbose = TRUE
)
}
\arguments{
\item{case}{Case in Coordinate class object format.}

\item{k}{Integer size of the expanded k-mer.}

\item{ctrl.rel.pos}{Control relative position.}

\item{genome}{Genome class object.}

\item{output.path}{Output directory path to save control coordinate.}

\item{verbose}{Boolean. Default is TRUE and will print progress updates.}
}
\value{
Control in Coordinate class object format.
}
\description{
Build control regions
}
