% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countChoppedKmers.R
\name{countChoppedKmers}
\alias{countChoppedKmers}
\title{Chop k-mers in given ranges of a sequence.}
\usage{
countChoppedKmers(dna.seq, starts, ends, k, method = "auto")
}
\arguments{
\item{dna.seq}{A string of sequence.}

\item{starts}{Start positions.}

\item{ends}{End positions.}

\item{k}{Size of kmer.}

\item{method}{Method: "Biostrings" or "substring". Default is Biostrings.}
}
\value{
A k-mer-named vector of counts.
}
\description{
Chop k-mers in given ranges of a sequence.
}
