% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countMidPatternContext2}
\alias{countMidPatternContext2}
\title{Locate a middle sequence pattern and count its sequence context.}
\usage{
countMidPatternContext2(sequence, mid_pattern, window, context_patterns)
}
\arguments{
\item{sequence}{A sequence to slide.}

\item{mid_pattern}{A middle pattern to search for.}

\item{window}{Size of a surrounding window.}

\item{context_pattern}{A context pattern to search for.}
}
\value{
A numeric named vector of frequencies with counts as their names.
}
\description{
Locate a middle sequence pattern and count its sequence context.
}
