% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countMidPatternKmers}
\alias{countMidPatternKmers}
\title{Count relavent k-mers with specified middle pattern from sequence string(s)
using a simple hash table.}
\usage{
countMidPatternKmers(sequences, k, mid_pattern)
}
\arguments{
\item{k}{K-mer size.}

\item{mid_pattern}{A middle pattern to search for.}

\item{sequence}{A sequence to slide.}
}
\value{
A k-mer-named vector of count.
}
\description{
Locate a middle sequence pattern and register and count k-mer.
}
