% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadUCSCgenome.R
\name{downloadUCSCgenome}
\alias{downloadUCSCgenome}
\title{Download chromosome-separated fasta genome sequence from UCSC database.}
\usage{
downloadUCSCgenome(genome.name, output.path, chr.name, method = "curl")
}
\arguments{
\item{genome.name}{Genome name e.g. hg19, hg38, mm19, etc.}

\item{output.path}{Output folder.}

\item{chr.name}{Specific chromosome to download. If not specified, default
is all.}

\item{method}{Download method for base::download.file function.}
}
\value{
An output folder containing chromosome-separated fasta files.
}
\description{
Download chromosome-separated fasta genome sequence from UCSC database.
}
