% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGenicElementCoor.R
\name{generateGenicElementCoor}
\alias{generateGenicElementCoor}
\title{Resolve and generate genic element coordinates from UCSC genePred table.}
\usage{
generateGenicElementCoor(
  genepred,
  element.names = "all",
  upstream = NULL,
  downstream = NULL,
  genome.name = NULL,
  genome = NULL,
  return.coor.obj = FALSE
)
}
\arguments{
\item{genepred}{UCSC genome name e.g. hg38 and mm39.}

\item{element.names}{Output genic element type: "all", "intron", "exon",
"CDS", or "UTR". Default is to output "all" type of genic element.names.}

\item{upstream}{Length of upstream sequence. It can overlap other gene.
Default is NULL i.e. not to output upstream.}

\item{downstream}{Length of downstream sequence. It can overlap other
gene. Default is NULL i.e. not to output downstream.}

\item{genome.name}{UCSC genome name. This is used to truncate overflowing
downstream and upstream coordinates. If genome is not supplied, trimming
will not be performed.}

\item{genome}{Genome object.}

\item{return.coor.obj}{Output Coordinate object? Default to FALSE.}
}
\value{
Genic element coordinates in `data.table` or `Coordinate` object.
}
\description{
Resolve and generate genic element coordinates from UCSC genePred table.
}
