% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIntergenicCoor.R
\name{generateIntergenicCoor}
\alias{generateIntergenicCoor}
\title{Resolve and generate genic element coordinates from UCSC genePred table.}
\usage{
generateIntergenicCoor(
  genepred,
  genome.name,
  igr.rel.pos = c(5000, 7500),
  igr.min.length = 150,
  return.coor.obj = FALSE
)
}
\arguments{
\item{genepred}{UCSC genePred table. Either the database used by UCSC to
generate the table ("refseq" or "gencode") or genePred \code{data.table} is
acceptable.}

\item{genome.name}{UCSC genome name e.g. hg38 and mm39.}

\item{igr.rel.pos}{Intergenic relative position. Default is c(5000, 7500)
i.e. from 5000 to 7500 bp from the genic regions.}

\item{igr.min.length}{Minimum length of the intergenic rregion. Default is
150.}

\item{return.coor.obj}{Return \code{Coordinate} object? Default is FALSE.}
}
\value{
Intergenic coordinates in \code{data.table}.
}
\description{
Resolve and generate genic element coordinates from UCSC genePred table.
}
