% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOSMICauthURL.R
\name{getCOSMICauthURL}
\alias{getCOSMICauthURL}
\title{Get COSMIC authenticated URL.}
\usage{
getCOSMICauthURL(email, password, url)
}
\arguments{
\item{email}{Email registered with COSMIC.}

\item{password}{Password.}

\item{url}{URL given by COSMIC website to access data.}
}
\value{
An authenticated URL, valid for 1 hour access.
}
\description{
To access the data for non-commercial usage, you must register with the
COSMIC. This function fetch the authenticated URL from the public URL
given by the COSMIC website.
}
