% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsemblData.R
\name{getEnsemblData}
\alias{getEnsemblData}
\title{A generic function to get Ensembl data persistently from a URL. This is an
   internal function used by other getEnsemblXXX functions.}
\usage{
getEnsemblData(url, handle, max.attempt = 5)
}
\arguments{
\item{url}{Pre-built Ensembl REST URL.}

\item{handle}{Curl handle object.}

\item{max.attempt}{Maximum number of attempt. Default to 5.}
}
\value{
A parsed json data. Could be data.frame or list of lists.
}
\description{
Error is handled based on their rule as set out at
https://github.com/Ensembl/ensembl-rest/wiki/HTTP-Response-Codes
}
