% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsemblRegionFeatures.R
\name{getEnsemblRegionFeatures}
\alias{getEnsemblRegionFeatures}
\title{Get features of a given region.}
\usage{
getEnsemblRegionFeatures(species, chromosome, start, end, features)
}
\arguments{
\item{species}{A species name or alias e.g. homo_sapiens, human, etc.}

\item{chromosome}{Ensembl chromosome name. Unlike UCSC, no prefix chr.}

\item{start}{Start position.}

\item{end}{End position.}

\item{features}{Region features. Options are "band", "gene", "transcript",
"cds", "exon", "repeat", "simple", "misc", "variation",
"somatic_variation", "structural_variation",
"somatic_structural_variation", "constrained", "regulatory", "motif",
"peak", "other_regulatory", "array_probe", and "mane".}
}
\value{
A `data.table` of Ensembl features.
}
\description{
GET overlap/region/:species/:region?feature=XXX?feature=XXXX?....
}
\details{
somatic_variation is from COSMIC
}
