% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGnomADvariants.R
\name{getGnomADvariants}
\alias{getGnomADvariants}
\title{Get gnomAD VCF file using tabix.}
\usage{
getGnomADvariants(
  chr.names,
  starts,
  ends,
  INFO.filter = NULL,
  version = "3.1.2",
  server = "random"
)
}
\arguments{
\item{chr.names}{Chromosome names.}

\item{starts}{Start positions.}

\item{ends}{End positions.}

\item{INFO.filter}{Parse only filtered INFO ID. Default is to parse all IDs.}

\item{version}{The gnomAD version. Default to latest version 3.1.2.}

\item{server}{Server locations: "google", "amazon", or "microsoft". Default
is random.}
}
\value{
A data.table of VCF.
}
\description{
Function retrieves variant data from gnomAD VCF files using tabix for a
specified set of genomic regions. It allows users to select the gnomAD
version and server location (Google, Amazon, or Microsoft) for fetching the
data.
}
