% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCBIassemblySummary.R
\name{getNCBIassemblySummary}
\alias{getNCBIassemblySummary}
\title{Get NCBI assembly summary.}
\usage{
getNCBIassemblySummary(organism.group, db = "refseq")
}
\arguments{
\item{organism.group}{A string specifying the taxonomic group for which the
assembly summary is requested. Options include 'archaea', 'bacteria', 'fungi',
'invertebrate', 'plant', 'protozoa', 'vertebrate_mammalian', 'vertebrate_other',
'viral', or 'all'.}

\item{db}{A string specifying the database to use, either 'refseq' or 'genbank'.}
}
\value{
A data.table containing the assembly summary for the specified taxonomic group.
}
\description{
Retrieves the assembly summary from NCBI for a specified taxonomic group.
This function allows users to obtain genome assembly information from either
RefSeq or GenBank databases for various taxonomic groups.
}
