% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUCSCgenePredTable.R
\name{getUCSCgenePredTable}
\alias{getUCSCgenePredTable}
\title{Study k-mer composition across species.}
\usage{
getUCSCgenePredTable(genome.name, db)
}
\arguments{
\item{db}{Database used by UCSC to generate the table: "refseq" or "gencode"}

\item{UCSC.genome.name}{UCSC genome name e.g. hg38 and mm39.}
}
\value{
genepred \code{data.table}.
}
\description{
Study k-mer composition across species.
}
