% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmertone.R
\name{kmertone}
\alias{kmertone}
\title{Kmertone all-in-one user interface}
\usage{
kmertone(
  case.coor.path,
  genome.name,
  strand.sensitive,
  k,
  ctrl.rel.pos = c(80, 500),
  case.pattern,
  output.dir = "output/",
  case,
  genome,
  control,
  control.path,
  genome.path,
  rm.case.kmer.overlaps,
  single.case.len,
  merge.replicates,
  kmer.table,
  module = "score",
  rm.dup = TRUE,
  case.coor.1st.idx = 1,
  ctrl.coor.1st.idx = 1,
  coor.load.limit = 1,
  genome.load.limit = 1,
  genome.fasta.style = "UCSC",
  genome.ncbi.db = "refseq",
  use.UCSC.chr.name = FALSE,
  verbose = TRUE,
  kmer.cutoff = 5,
  selected.extremophiles,
  other.extremophiles,
  cosmic.username,
  cosmic.password,
  tumour.type.regex = NULL,
  tumour.type.exact = NULL,
  cell.type = "somatic",
  genic.elements.counts.dt,
  population.size = 1e+06,
  selected.genes,
  add.to.existing.population = FALSE,
  population.snv.dt = NULL,
  pop.plot = TRUE,
  pop.loop.chr = FALSE
)
}
\description{
Kmertone all-in-one user interface
}
\section{Fields}{

\describe{
\item{\code{case.coor.path}}{A path to a folder containing either (1) chromosome-
separated coordinate files (assume replicates for subfolder) or (2)
bedfile. (assume replicates for bedfiles).}

\item{\code{genome.name}}{Name of genome. Kmertone include UCSC genome e.g. "hg19"
and "hg38". Default is "unknown".}

\item{\code{strand.sensitive}}{Does strand polarity matters? Default is TRUE. In
sensitive strand, case k-mers are extracted from the sense strand only.
In contrast, in insensitive strand, the case k-mers are extracted from
both strands.}

\item{\code{k}}{Length of k-mer to be investigated. Recommended is 7 or 8.}

\item{\code{ctrl.rel.pos}}{A relative range position of control regions. For example
c(80,500) means control regions are 80-500 bases away from the the case
site (upstream and downstream).}

\item{\code{case.pattern}}{Case pattern(s). Default is NULL for no pattern. Only
applicable to module other than study_cancer_genes.}

\item{\code{output.dir}}{An output directory name. Default is output/.}

\item{\code{case}}{Optional pre-built Coordinate object. Option for
strand.sensitive and single.case.len will be ignored.}

\item{\code{genome}}{Optional pre-built Genome object. Option genome.name and
genome.path will be ignored.}

\item{\code{control}}{Optional pre-built control Coordinate object. Option
ctrl.rel.pos will be ignored.}

\item{\code{control.path}}{A path for pre-built control Coordinate object. Option
ctrl.rel.pos will be ignored.}

\item{\code{genome.path}}{A path to a directory of user own genome. The directory
must contain chromosome-separated fasta files. The name of the fasta
files must be similar to chromosome name in the coordinate table.}

\item{\code{rm.case.kmer.overlaps}}{Remove overlapping case k-mers? This is useful
to completely obliterate potential neighboring effects.}

\item{\code{single.case.len}}{If the case is in uniform length, only start positions
are considered. This is to remove redundancy of the coordinates and
reduce memory usage.}

\item{\code{merge.replicates}}{Merge the replicates if exist? Merging will remove
all duplicate positions from multiple replicates, resulting in a single
unique individual point. Not merging replicate will result in weighted
k-mer counts. Default is TRUE.}

\item{\code{kmer.table}}{K-mer table with pre-calculated scores. It can be directly
used in other kmertone module.}

\item{\code{module}}{Kmertone module. Options are: "score" (calculate z-score),
"explore" (perform exploratory analysis), "tune" (find the best length
of k-mer to be used).}

\item{\code{rm.dup}}{Remove duplicate coordinate? Remove duplicates within replicate
e.g. sequencing read duplicates. Default is TRUE.}

\item{\code{case.coor.1st.idx}}{Case coordinate indexing format whether
zero-based [start, end) or one-based \link{start, end}. Default is 1.}

\item{\code{ctrl.coor.1st.idx}}{Control coordinate indexing format whether
zero-based [start, end) or one-based \link{start, end}. Default is 1.}

\item{\code{coor.load.limit}}{Maximum loaded chromosome coordinate data. Default is
1.}

\item{\code{genome.load.limit}}{Maximum loaded chromosome sequence data. Default is
1.}

\item{\code{genome.fasta.style}}{Genome FASTA style: "UCSC" or "NCBI". Default
is "UCSC"}

\item{\code{genome.ncbi.db}}{For NCBI Genome, select database to use: "refseq" or
"genbank". Default is "refseq".}

\item{\code{use.UCSC.chr.name}}{For NCBI Genome, use UCSC chromosome name?}

\item{\code{verbose}}{Print message? Default is TRUE.}

\item{\code{kmer.cutoff}}{Percent kmer cutoff for the case studies. Default is 5.}

\item{\code{selected.extremophiles}}{A vector of selected extremophile species. e.g.
c("Deinococcus soli", "Deinococcus deserti")
The best representative will be selected from the assembly summary.}

\item{\code{other.extremophiles}}{A vector of other extremophile species. These are
used as a control to compare with the selected extremophiles.}

\item{\code{cosmic.username}}{COSMIC username (email). This is used to get cancer
gene census.}

\item{\code{cosmic.password}}{COSMIC password.}

\item{\code{tumour.type.regex}}{Tumour-type regular expression to filter the cancer
gene census table.}

\item{\code{tumour.type.exact}}{Exact tumour type to be included in the cancer gene
census table.}

\item{\code{cell.type}}{Cell type to be included in the cancer gene census table.
Default is somatic cell.}

\item{\code{genic.elements.counts.dt}}{Counts of susceptible k-mers in the genic
elements, generated from STUDY_GENIC_ELEMENTS.}

\item{\code{population.size}}{Size of population in STUDY_ACROSS_POPULATION. Default
is 1 million.}

\item{\code{selected.genes}}{Selected genes to be mutated based on SNVs in
STUDY_ACROSS_POPULATION.}

\item{\code{add.to.existing.population}}{STUDY_ACROSS_POPULATION can be run multiple
times to add to the existing simulated individuals. Default is FALSE to
overwrite.}

\item{\code{population.snv.dt}}{SNVs used in the population simulation.}

\item{\code{pop.plot}}{To plot the outcome of STUDY_ACROSS_POPULATION. Default is
TRUE.}

\item{\code{pop.loop.chr}}{In STUDY_ACROSS_POPULATION, should the operation loop
based on chromosome name? Default is FALSE.}
}}

