% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapKmers.R
\name{mapKmers}
\alias{mapKmers}
\title{Map k-mers of a given sequence and coordinate.}
\usage{
mapKmers(seq, start, end = NULL, len = NULL, k, rm.trunc.kmer = TRUE)
}
\value{
Mapped k-mers.
}
\description{
Map k-mers of a given sequence and coordinate.
}
\section{Fields}{

\describe{
\item{\code{seq}}{A single sequence string.}

\item{\code{start}}{A single of vector of start coordinates where k-mers are mapped.}

\item{\code{end}}{A single of vector of end coordinates where k-mers are mapped. It
is NULL when all regions have the same length.}

\item{\code{len}}{A single integer for specifying fixed length of regions. The end
is NULL in this case.}

\item{\code{k}}{Length of k-mer to be mapped.}

\item{\code{rm.trunc.kmer}}{Remove truncated k-mers as a result of out-of-bound
region. Default is TRUE.}
}}

