% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionCoordinate.R
\name{partitionCoordinate}
\alias{partitionCoordinate}
\title{Partition overlapping or continuous regions.}
\usage{
partitionCoordinate(coor)
}
\value{
Partitioned coordinate `data.table`.
}
\description{
Table must have start and end columns. The mechanism is similar to the
   disjoin function from GenomicRanges but the end coordinate is different.
}
\section{Fields}{

\describe{
\item{\code{coor}}{Coordinate `data.table`.}
}}

