% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBED.R, R/writeBED.R
\name{readBED}
\alias{readBED}
\title{Read a BED file. One-based indexing is enforced.}
\usage{
readBED(bed, output.filename)

readBED(bed, output.filename)
}
\arguments{
\item{bed}{A BED `data.table`.}

\item{output.filename}{An output BED filename.}

\item{bed.path}{A path to a BED file.}
}
\value{
A `data.table` of BED.
}
\description{
Read a BED file. One-based indexing is enforced.

Write a BED file. Zero-based indexing is enforced.
}
