% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGenomesForCrossSpeciesStudy.R
\name{selectGenomesForCrossSpeciesStudy}
\alias{selectGenomesForCrossSpeciesStudy}
\title{Select genomes for cross-species studies.}
\usage{
selectGenomesForCrossSpeciesStudy(organism.group = "bacteria", db = "refseq")
}
\arguments{
\item{organism.group}{Species group: archaea, bacteria, fungi, invertebrate,
plant, protozoa, vertebrate_mammalian, vertebrate_other, or viral.}

\item{db}{Database record to use: refseq or genbank}
}
\value{
NCBI assembly summary with added column organism.group.
}
\description{
The following filters are applied:
\enumerate{
\item assembly_level: "Complete Genome" or "Chromosome"
\item genome_rep: "Full"
\item Unique species_taxid (single representative species)
\item refseq_category of "reference genome" is prioritised over
"representative genome"
}
}
