% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{splitFASTA2}
\alias{splitFASTA2}
\title{Read fast single-header FASTA file.}
\usage{
splitFASTA2(file_path, output_dir)
}
\arguments{
\item{file_path}{A path to FASTA file.}

\item{unmask}{Capitalize all base letters?}
}
\value{
A single sequence string without header.
}
\description{
Slower than data.table::fread
}
