% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimCoordinate.R
\name{trimCoordinate}
\alias{trimCoordinate}
\title{Trim out-of-bound coordinates}
\usage{
trimCoordinate(coor, seq.len, genome)
}
\arguments{
\item{coor}{Coordinate \code{data.table}.}

\item{seq.len}{Sequence length to trim end position.}

\item{genome}{\code{Genome} class object.}
}
\value{
Trimmed coordinate \code{data.table}.
}
\description{
It operates in two mode: coordinate table with and without chromosome. The
former require \code{Genome} for the chromosomal sequence length.
}
